requirements.txt
torch>=2.2.0
torchvision>=0.17.0
tqdm>=4.66
scikit-learn>=1.2
matplotlib>=3.7
pillow>=10.0
numpy>=1.24,<2
pyyaml>=6.0
